add: speed calculator, profile security

+++ src/Controller/RunController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\Request;
 
@@ -15,13 +16,17 @@ class RunController extends Controller
     /**
      * @Route("/run/delete_run/{id}", name="delete_run")
      */
-     public function new($id, Request $request)
+     public function new($id, Request $request, AuthorizationCheckerInterface $authChecker)
      {
           $em = $this->getDoctrine()->getManager();
 
           $run = $this->getDoctrine()
               ->getRepository(Run::class)
               ->find($id);
+
+          if ($authChecker->isGranted('delete', $run) !== true) {
+               return $this->redirectToRoute('show_user', array('id' => $run->getUser()->getId()));
+          }
 
           $em->remove($run);
           $em->flush();
